requirements.txt
# Core Django Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Data Analysis & Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# API & Data Integration
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Visualization & Charts
plotly>=5.18.0

# Database & Caching
psycopg2-binary>=2.9.7
redis>=5.0.0
django-redis>=5.3.0

# Production & Deployment
gunicorn>=21.2.0
whitenoise>=6.5.0

# Monitoring & Error Tracking
sentry-sdk[django]>=1.32.0 